+++ psygnal/_evented_model.py
